requirements.txt
atomicwrites==1.4.0
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
colorama==0.4.4
Django==2.2.16
django-debug-toolbar==3.2.4
Faker==12.0.1
flake8==4.0.1
flake8-broken-line==0.4.0
flake8-isort==4.1.1
flake8-plugin-utils==1.3.2
flake8-polyfill==1.0.2
flake8-return==1.1.3
idna==3.3
iniconfig==1.1.1
isort==5.10.1
mccabe==0.6.1
mixer==7.1.2
packaging==21.3
pep8-naming==0.12.1
Pillow==9.0.1
pluggy==0.13.1
py==1.11.0
pycodestyle==2.8.0
pyflakes==2.4.0
pyparsing==3.0.7
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dateutil==2.8.2
pytz==2022.1
requests==2.26.0
six==1.16.0
sorl-thumbnail==12.7.0
sqlparse==0.4.2
testfixtures==6.18.5
toml==0.10.2
urllib3==1.26.9
